requirements.txt
torch
torchvision
gymnasium[atari,accept-rom-license]
ale-py
opencv-python
numpy
pandas
Pillow
tqdm
